+++ tensorflow/compiler/xla/mlir_hlo/thlo/transforms/legalize_sort/legalize_sort.cc
@@ -371,11 +371,11 @@ struct Slicer {
         strides(inductionVariables.size() + 1, b.getI64IntegerAttr(1)) {
     sizes[sortDim] = sortDimSize;
     for (size_t i = 0; i < inductionVariables.size() + 1; ++i) {
-      if (i == sortDim) {
+      if ((int64_t)i == sortDim) {
         offsets.push_back(b.getI64IntegerAttr(0));
       } else {
         offsets.push_back(
-            inductionVariables[i - static_cast<int>(i > sortDim)]);
+            inductionVariables[i - static_cast<int>((int64_t)i > sortDim)]);
       }
     }
   }
